==> code/01_scrape_wiki_state_list.ipynb <==
"""

Get list of state pages from the Wikipedia page
https://en.wikipedia.org/wiki/County_(United_States)

scrape page and output to pickle the state name and wikipedia link

scrape page and output to pickle the table with county aggregates by state

"""

from bs4 import BeautifulSoup
import requests

import numpy as np
import pandas as pd

import pickle

url = 'https://en.wikipedia.org/wiki/County_(United_States)' 

response = requests.get(url)

response.status_code  #200 = success!

response.text[:1000]  #First 1000 characters of the HTML

page = response.text

soup = BeautifulSoup(page)

#soup

#print(soup.prettify())

table = soup.find_all(class_ ='flagicon')
#table


headers = ['state', 'wiki_link']

state_pages = []

for st in table:

    state = st.find('a').get('title')
    state_link = st.find('a').findNext('a').get('href')

    links_dict = dict(zip(headers, [state, state_link]))

    state_pages.append(links_dict)


#state_pages

# put the list of links into a df
df_state_links = pd.DataFrame(state_pages)

df_state_links.head()

len(df_state_links)

df_state_links.tail(10)

with open('state_wiki_links.pickle', 'wb') as to_write:
    pickle.dump(df_state_links, to_write)

# for each state and territory listed scrape from page table:

# state
# 2019 total population
# land area
# counties
# equivalents
# total counties + equivalents
# avg population 
# avg land area 

table_data = soup.find(class_='wikitable sortable').find_all('td')
#table_data

state_data = [x.text.strip() for x in table_data]
#state_data

len(state_data)

state_data_array = np.array(state_data).reshape(59,8)
#state_data_array

df_st_data = pd.DataFrame(state_data_array)
df_st_data.head(5)

col_titles = ['State, federal district, or territory',
          'Total 2019 population',
          'Total Land area',
          'Counties',
          'Equivalents',
          'Total',
          'Average Population',
          'Average Land area']

df_st_data.columns = col_titles

df_st_data.head(5)

with open('state_wiki_data.pickle', 'wb') as to_write:
    pickle.dump(df_st_data, to_write)
