==> code/02_scrape_wiki_county_list.ipynb <==
"""

Get list of each state's county pages 
from each of the state Wikipedia pages
scraped with code in '01_scrape_wiki_state_list'


output to pickle

"""

from bs4 import BeautifulSoup
import requests

import numpy as np
import pandas as pd

import pickle

import time

# list of state page wikipedia links 

with open('state_wiki_links.pickle', 'rb') as read_file:
    state_links = pickle.load(read_file)
    

state_links.head()

st_dc_only = state_links[ :-6]

#drop dc as dc is itself and does not have any sub-state/county areas

st_only = st_dc_only[~st_dc_only['state'].isin(['Washington, D.C.'])].reset_index(drop=True)


st_only.head(5)

base_url = 'https://en.wikipedia.org'

county_pages = []

for i,state in enumerate(st_only['state']):
    
    state_page_url = base_url + st_only['wiki_link'][i]
    
    response = requests.get(state_page_url)
    page = response.text
    soup = BeautifulSoup(page)
    
    table_data = soup.find(class_='wikitable sortable').find_all('th' , scope='row')
    
    headers = ['state', 'FIPS_code', 'county_name', 'wiki_link']
    state = st_only['state'][i]

    for result in table_data:  
        try:
                        
            FIPS_code = result.findNext('td').text.strip()
            county_name = result.findNext('a').get('title')
            wiki_link = result.findNext('a').get('href') 
            
            
            links_dict = dict(zip(headers, [state, FIPS_code, county_name, wiki_link]))

            county_pages.append(links_dict)
            
        except:
            print('ERROR: ' , state_page_url)
            continue
            
    time.sleep(1)


#county_pages

county_links_df = pd.DataFrame(county_pages)

county_links_df.head(5)

len(county_links_df)

county_links_df.to_csv('review_county_links.csv')
# reviewed table in CSV

# drop row from Washington state page that is linking to census.gov for info about all states

all_county_pages = county_links_df[~county_links_df['wiki_link'].str.contains('census.gov')].reset_index(drop=True)

len(all_county_pages)

with open('county_wiki_links.pickle', 'wb') as to_write:
    pickle.dump(all_county_pages, to_write)
