==> code/03_scrape_wiki_county_page.ipynb <==
"""

Get data for features from each of the US county Wikipedia pages
Loop over list of URLs of Wikipedia pages for each US county 

features to collect:
major highways
city or cities
town or towns

output to pickle

Run time: 1 hour

"""

from bs4 import BeautifulSoup
import requests

import numpy as np
import pandas as pd

import pickle
import time

import regex as re

from datetime import datetime 

# list of county pages wikipedia links 

with open('county_wiki_links.pickle', 'rb') as read_file:
    county_links = pickle.load(read_file)

county_links.head(5)

len(county_links)

base_url = 'https://en.wikipedia.org'


now = datetime.now () 
current_time = now.strftime ("%H:%M:%S") 
print("Current Time =", current_time)

cnty_data = []

for i, county in enumerate(county_links['county_name']):

    county_page_url = base_url + county_links['wiki_link'][i]

    response = requests.get(county_page_url)
    page = response.text
    soup = BeautifulSoup(page)
        
    state = county_links['state'][i]
    county_name = county_links['county_name'][i]
    FIPS_code = county_links['FIPS_code'][i]
    
    headers = ['state', 'county_name', 'FIPS_code', 'county_page_url', 'highways', 'hwy_count',
                    'cities', 'city_count', 'towns', 'town_count']

    try:

        try:
            highway_find = soup.find(id=re.compile('highway')).findNext('ul').find_all('a')              
            highway_list = [x.text for x in highway_find]
            highways = list(filter(len, highway_list))  # remove blanks for the hwy sign images
            hwy_count = len(highways)

        except:
            highways = []
            hwy_count = 0

        try:           
            city_list = soup.find(id=re.compile(r"(\bCity\b|\bCities\b)")).findNext('ul').find_all('li')
            cities = [x.find('a').text for x in city_list]
            city_count = len(cities)

        except:
            cities = []
            cities_count = 0

        try:
            town_list = soup.find(id=re.compile(r"(\bTown\b|\bTowns\b)")).findNext('ul').find_all('li')
            towns = [x.find('a').text for x in town_list]
            town_count= len(towns)

        except:
            towns = []
            town_count = 0
            
        
        features_dict = dict(zip(headers, [state, county_name, FIPS_code, county_page_url, highways, 
                                           hwy_count, cities, city_count, towns, town_count]))
        
        cnty_data.append(features_dict)
                

    except:
        print('ERROR with URL: ', county_page_url)
        continue


    time.sleep(1)
        


now = datetime.now () 
current_time = now.strftime ("%H:%M:%S") 
print("Current Time =", current_time)

len(cnty_data)

# to df
county_features = pd.DataFrame(cnty_data)


len(county_features)

county_features.head()

# export all and review
county_features.to_csv('review_features_scrape.csv', index=False)

# output pickle
with open('county_features.pickle', 'wb') as to_write:
    pickle.dump(county_features, to_write)
