==> code/04_scrape_FIPS_codes.ipynb <==
"""

Get list of state FIPS codes from the USDA page
https://www.nrcs.usda.gov/wps/portal/nrcs/detail/?cid=nrcs143_013696

scrape page and output to pickle


"""

from bs4 import BeautifulSoup
import requests

import numpy as np
import pandas as pd

import pickle

url = 'https://www.nrcs.usda.gov/wps/portal/nrcs/detail/?cid=nrcs143_013696' 

response = requests.get(url)

response.status_code  #200 = success!

response.text[:1000]  #First 1000 characters of the HTML

page = response.text

soup = BeautifulSoup(page)

#soup

#print(soup.prettify())

table = soup.find(id ='detail').find_all('td')
#table


fips_table_data = [x.text.strip() for x in table]
#fips_table_data

data_array = np.array(fips_table_data).reshape(55, 3)

#data_array

st_fips_codes = pd.DataFrame(data_array)
st_fips_codes.columns = ['state', 'abbr', 'FIPS']

st_fips_codes.head()

with open('st_fips_codes.pickle', 'wb') as to_write:
    pickle.dump(st_fips_codes, to_write)
