==> code/05_usda_unemp_inc_file.ipynb <==
"""
Filter data file with unemployment data and median household income data from USDA to year 2019.
Pivot table from tall to wide.

Output to pickle
"""

import pandas as pd
import numpy as np
import pickle


unemp_inc_all = pd.read_csv('../data/USDA_unemp_inc.csv')
unemp_inc_all.shape

unemp_inc_all.info()

unemp_inc_all.head()

unemp_inc_all.Attribute.value_counts()

unemp_inc_all.set_index(['Attribute'], inplace=True)
unemp_inc_all.head()

#  don't need 2000 data elements


unemp_inc_2019 = unemp_inc_all.drop(['Employed_2000', 'Civilian_labor_force_2000', 
                                     'Unemployed_2000', 'Unemployment_rate_2000'], 
                                    axis=0).reset_index()

unemp_inc_2019.head()

unemp_inc_2019.Attribute.value_counts()

unemp_inc_2019['area_name'].value_counts()

unemp_inc_2019[unemp_inc_2019['area_name'] == 'District of Columbia']

unemp_inc_2019[unemp_inc_2019['area_name'] == 'Lee County, NC']

unemp_inc_2019[unemp_inc_2019['area_name'] == 'Autauga County, AL']

unemp_inc_pivot = pd.pivot_table(unemp_inc_2019,
                                 values='Value', 
                                 columns='Attribute', 
                                 index=['fips_txt','Stabr', 'area_name']).reset_index(drop=False)

unemp_inc_pivot.head()

with open('unemp_inc_pivot.pickle', 'wb') as to_write:
    pickle.dump(unemp_inc_pivot, to_write)
