==> code/06_usda_pop_est_file.ipynb <==
"""
Filter data file with population data from USDA to year 2019.

Output to pickle
"""

import pandas as pd
import numpy as np
import pickle


pop_est_all = pd.read_csv('../data/USDA_PopulationEstimates.csv', encoding='unicode-escape')
pop_est_all.shape

pop_est_all.info()

pop_est_all.head()

pop_est_all.columns

pop_est_filter = pop_est_all[['FIPStxt', 'State', 'Area_Name', 'POP_ESTIMATE_2019', 'N_POP_CHG_2019']]
pop_est_filter.head()

with open('pop_est_2019.pickle', 'wb') as to_write:
    pickle.dump(pop_est_filter, to_write)
