==> code/07_usda_typology_file.ipynb <==
"""
Import USDA file with county typology data

Output to pickle
"""

import pandas as pd
import numpy as np
import pickle


cnty_typol = pd.read_csv('../data/2015CountyTypologyCodes.csv', encoding='unicode-escape')
cnty_typol.shape

cnty_typol.info()

cnty_typol.head(10)

cnty_typol.columns

with open('county_typology.pickle', 'wb') as to_write:
    pickle.dump(cnty_typol, to_write)
