==> code/08_combine_all_data.ipynb <==
"""
combine data gathered and saved in several pickle files

output to one combined pickle file
"""

import pickle
import pandas as pd
import numpy as np

# state FIPS codes
with open('st_fips_codes.pickle', 'rb') as read_file:
    fips_codes = pickle.load(read_file)
    
fips_codes.head()

# features from Wikipedia
with open('county_features.pickle', 'rb') as read_file:
    county_features = pickle.load(read_file)
    
county_features.head()

county_features.dtypes

# add state fips code to county features df
# check state names exactly the same

county_features['state'].unique()

fips_codes['state'].unique()

# revise state names in the county features df to remove specification of (state)
county_features.loc[county_features['state'] == 'Georgia (U.S. state)', ['state']] = 'Georgia'
county_features.loc[county_features['state'] == 'New York (state)', ['state']] = 'New York'
county_features.loc[county_features['state'] == 'Washington (state)', ['state']] = 'Washington'


county_features['state'].unique()

cnty_features = pd.merge(county_features, fips_codes, how='left', on='state' )
cnty_features.head()

cnty_features.tail()

# import pickle files with usda data

with open('pop_est_2019.pickle', 'rb') as read_file:
    pop_est = pickle.load(read_file)
    
    
pop_est.head()

pop_est.dtypes

pop_est['FIPStxt'] = pop_est['FIPStxt'].astype(str)
pop_est.dtypes

# check state rows have FIPS ending in '000'

pop_est[pop_est['FIPStxt'].str[-3:] == '000']

# keep only counties in pop_est df

cnty_pop_est = pop_est[pop_est['FIPStxt'].str[-3:] != '000']

cnty_pop_est.head()

# remove all US row
cnty_pop_est = cnty_pop_est[cnty_pop_est['State'] != 'US']

cnty_pop_est.head()

cnty_pop_est['FIPS_code'] = cnty_pop_est['FIPStxt'].apply(lambda x: str(x).zfill(5))

cnty_pop_est.head()

len(cnty_pop_est)

with open('unemp_inc_pivot.pickle', 'rb') as read_file:
    unemp_inc = pickle.load(read_file)
    
unemp_inc.head()

unemp_inc.dtypes

unemp_inc['fips_txt'] = unemp_inc['fips_txt'].astype(str)
unemp_inc.dtypes

# keep only counties
cnty_unemp_inc = unemp_inc[unemp_inc['fips_txt'].str[-3:] != '000']
cnty_unemp_inc.head()

cnty_unemp_inc = cnty_unemp_inc[cnty_unemp_inc['Stabr'] != 'US']

len(cnty_unemp_inc)

# fips_txt needs leading 0 for states 1-9

cnty_unemp_inc['FIPS_code'] = cnty_unemp_inc['fips_txt'].apply(lambda x: str(x).zfill(5))

cnty_unemp_inc.head()

# new column to combine state FIPS and county FIPS in cnty_features df

cnty_features['FIPS_code_NEW'] = cnty_features['FIPS'] + cnty_features['FIPS_code']
cnty_features.head()

with open('county_typology.pickle', 'rb') as read_file:
    cnty_type = pickle.load(read_file)
    
cnty_type.head()

len(cnty_type)

cnty_type['FIPS_code'] = cnty_type['FIPStxt'].apply(lambda x: str(x).zfill(5))

cnty_type.head()

# merge cnty_unemp_inc, cnty_pop_est, cnty_type, cnty_features


all_data_1 = pd.merge(cnty_unemp_inc, cnty_pop_est, how='left', on='FIPS_code')

all_data_2 = pd.merge(all_data_1, cnty_type, how='left', on='FIPS_code')

all_data_3 = pd.merge(all_data_2, cnty_features, how='left', left_on='FIPS_code', right_on='FIPS_code_NEW')

all_data_3.head()

all_data_3['Stabr'].unique()

# drop PR

county_data = all_data_3[all_data_3['Stabr'] != 'PR']

county_data['Stabr'].unique()

len(county_data)

#check results of merge
county_data.to_csv('county_data_all.csv', index=False)

with open('county_data.pickle', 'wb') as to_write:
    pickle.dump(county_data, to_write)
